Fix error message and column field not being validated for histogram.
